Use better string formatting
By using C-style string formatting, debug messages are only constructed
if the logger actually routes the message somewhere. This is only the
slightest of performance tweaks, but it's good to be consistent and use
proper idiomatic Python.

Elsewhere (where strings are constructed and immediately used), make use
of f-strings. They're more concise and readable!