Handle case of RSS link being missing
Occasionally, there may not be a link to the full record of a particular
item in the catalog (searching Seattle without narrowing it down to a
branch caused this error on my current GoodReads shelf).

Since this error only affects a few books in the shelf, raise a warning
and continue execution.