Use serverless to automatically deploy to Lambda
Initial testing of Lambda functions can be done by manually copying the
contents of a virtualenv to a `.zip` archive, uploading to AWS, and
configuring via the GUI.

However, controlling infrastructure via code is much more sensible!
Serverless is a good framework for meeting our needs.

We choose to Dockerize the pip environment because both grequests and
lxml are C-based, and need to compile binaries for use on the Lambda VM.
By using Docker, we can make use of an image that mirrors exactly what
AWS will run in Lambda-land.

I should add some mention of this to the README...