Restrict results to English
I keep a number of classics on my reading list, for which there are many
translations into other languages. So that the results aren't filled
with duplicate copies of the same book, search only for English.

Books that I want to read in other languages generally have English
translations, so this is a good reminder that they're on my list (and
that I should instead try querying with that specific language instead).