Add DynamoDB table for caching results
The API query to request books on a user's Goodreads shelf takes
approximately five seconds. It also represents a dataset that's unlikely
to change terribly often. In short, this query is an *excellent*
candidate for a caching mechanism.

I prefer relational databases, but DynamoDB provides a simple, on-demand
key-value store that we can use to cache results. It also includes
automatic record invalidation which should prove useful for adhering to
Goodreads' request that data not be cached longer than 24 hours.