Return if shelf results came from the cache
We recently implemented a simple caching mechanism, but it was opaque to
users whether or not results were just obtained, or came from the cache.
Report a simple boolean indicating cache hit/miss, plus a timestamp of
when the cache was last updated.

Callers can use this information to decide if they want to request new
results, or just to display when the data was last fetched.